+++ src/core/services/user/user-impl.service.ts
@@ -1,8 +1,8 @@
 import { AxiosResponse } from "axios";
 import { map, Observable } from "rxjs";
 
+import { UserDomain } from "@/core/entities/domains/user.domain";
 import { HttpClient } from "@/core/libs/http/http-client";
-import { UserDomain } from "@/domain/user.domain";
 
 import { UserService } from "./user.service";
