+++ src/core/usecases/header/header-impl.usecase.ts
@@ -1,7 +1,7 @@
 import { map, Observable } from "rxjs";
 
-import { UserService } from "@/services/user/user.service";
-import { HeaderPresenter } from "@/presenters/header.presenter";
+import { HeaderPresenter } from "@/core/entities/presenters/header.presenter";
+import { UserService } from "@/core/services/user/user.service";
 
 import { HeaderUsecase } from "./header.usecase";
