+++ src/core/usecases/header/header-mock.usecase.ts
@@ -1,8 +1,8 @@
 import { BehaviorSubject, Observable } from "rxjs";
 
-import { UserDomain } from "@/domain/user.domain";
-import { UserService } from "@/services/user/user.service";
-import { HeaderPresenter } from "@/presenters/header.presenter";
+import { UserDomain } from "@/core/entities/domains/user.domain";
+import { HeaderPresenter } from "@/core/entities/presenters/header.presenter";
+import { UserService } from "@/core/services/user/user.service";
 
 import { HeaderUsecase } from "./header.usecase";
